src/requirements.txt
boto3==1.34.0
openai==1.6.1
supabase==2.7.4
telnyx==2.1.5 # 2.0.0 errors on setup
python-dotenv==1.0.0
google-auth==2.23.4
google-api-python-client==2.108.0
